refactor: move analytics interval to seperate file
