+++ src/components/analytics-control-bar/index.tsx
@@ -22,6 +22,7 @@ import {
   AddButton,
   FilterDeleteButton,
 } from './utilities';
+import AnalyticsIntervalSelector, { AnalyticsInterval } from './interval';
 
 import {
   AppBar,
@@ -448,46 +449,6 @@ export function AnalyticsSpaceSelector(props: AnalyticsSpaceSelectorProps) {
   );
 }
 
-
-export enum AnalyticsInterval {
-  DAY = 'DAY',
-  HOUR = 'HOUR',
-  FIFTEEN_MINUTES = 'FIFTEEN_MINUTES',
-}
-
-const INTERVAL_CHOICES = [
-  { id: AnalyticsInterval.DAY, label: 'Day' },
-  { id: AnalyticsInterval.HOUR, label: 'Hour' },
-  { id: AnalyticsInterval.FIFTEEN_MINUTES, label: '15 Minutes' },
-];
-
-export const AnalyticsIntervalSelector = ({ value, onChange }) => {
-  const [ open, setOpen ] = useState(false);
-
-  const choice = INTERVAL_CHOICES.find(choice => choice.id === value);
-
-  return (
-    <div className={styles.analyticsIntervalSelector}>
-      <span className={styles.analyticsIntervalLabel}>by</span>
-      <Filter
-        open={open}
-        onOpen={() => setOpen(true)}
-        onClose={() => setOpen(false)}
-        text={<FilterBold>{choice ? choice.label : '(unknown interval)'}</FilterBold>}
-      >
-        <ItemList
-          choices={INTERVAL_CHOICES}
-          onClick={choice => {
-            onChange(choice.id);
-            // Blur the element that was selected if there was a focused element
-            if (document.activeElement) { (document.activeElement as HTMLElement).blur(); }
-            setOpen(false);
-          }}
-        />
-      </Filter>
-    </div>
-  );
-};
 
 
 enum RangeType {
